backend/src/requirements.txt
fastapi==0.120.1
uvicorn[standard]==0.34.0
httpx==0.28.1
coverage~=7.3.1
supabase==2.13.0
python-dotenv==1.0.1
python-multipart